Add support for custom clock renderers
* Clock renderers become stateless
* Clock renderer of TimeRangePicker can be easily changed to custom
* Default implementation of ClockRenderer added
